# reduce fallback scope