@wulianglongrd 2023-11-18
To clarify: the benchmark data above is overly optimistic, showing -47.64% reduction for `middle-exact`(sec/op) and -62.78% for `big-exact`. We applied this improvement to the production environment, and after a period of observation, the actual average time consumption of [this indicator](https://github.com/istio/istio/blob/235f899c9464bde5605144e5a1d4da3ebdb56cb4/pilot/pkg/xds/monitoring.go#L105) was reduced by about 30% (please note: our production environment has not yet used wildcard hosts)